pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "django_lightrag"
version = "0.1.0"
description = "LightRAG implementation as a reusable Django app"
authors = [{name = "LightRAG Team", email = "team@lightrag.ai"}]
license = "MIT"
readme = "README.md"
requires-python = ">=3.13"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 5.2",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    "Django>=5.2.8",
    "django-ninja>=1.0",
    "real-ladybug>=0.1.0",
    "chromadb>=0.4.0",
    "tiktoken>=0.5.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.0.0",
    "numpy>=1.24.0",
    "asyncio-throttle>=1.0.2",
    "pytest-django>=4.11.1",
    "django-llm-chat",
    "embed-gen",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-django>=4.5.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "django-stubs>=4.2.0",
]

[project.urls]
Homepage = "https://github.com/HKUDS/LightRAG"
Repository = "https://github.com/HKUDS/LightRAG"
Documentation = "https://docs.lightrag.dev/"

[tool.setuptools.packages.find]
include = ["django_lightrag*"]

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "DJ", "PIE", "PT", "RSE"]
ignore = ["E501", "PIE796"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "django_lightrag.tests.settings"
python_files = "test_*.py"
pythonpath = "."
markers = [
    "live_e2e: marks tests as live end-to-end tests (requires real services)",
]

[tool.uv.sources]
django-llm-chat = { path = "../llm_chat", editable = true }
embed-gen = { path = "../embed_gen", editable = true }

[dependency-groups]
dev = [
    "flake8>=5.0.4",
    "model-bakery>=1.23.3",
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
    "ruff>=0.15.6",
]
